Create README.md
Because that iz a big file, I devide it into small zip files. The input data for this project is on the the link https://www.kaggle.com/c/home-credit-default-risk.
I changed the name of input file excel following the image.png. We will import this into the next step below
2.create_test_data.ipynb is used for create out of sample dataset which is the same as what I do with training dataset. 
9.test_data.xlsx is the result of the previous step. 
all_process_code.rar having working_file_python.ipynb in is the most important file which is all the process. 
final_destination.xlsx is the result for prediction out of sample.